Basic Matrix4 changes

+++ include/math/matrix/matrix3.h
@@ -2,6 +2,6 @@
 #define MATRIX3_H
 
 //3x3 matrix represented as a 2d array
-typedef double matrix3 [3][3];
+typedef float matrix3 [3][3];
 
 #endif
